:sparkles: feat: 운영 날짜와 시간 판별 기능 구현
