+++ README.md
@@ -126,8 +126,8 @@
 - [x] Csv 형식의 파일 로딩 및 처리
 
 ### Evaluator
-- [ ] 운영 날짜 판별
-  - [ ] 주말 날짜 판별
-  - [ ] 공휴일 날짜 판별
-- [ ] 운영 시간 판별
+- [x] 운영 날짜 판별
+  - [x] 주말 날짜 판별
+  - [x] 공휴일 날짜 판별
+- [x] 운영 시간 판별
