+++ src/main/java/controller/AttendanceController.java
@@ -12,11 +12,11 @@ import domain.menu.Menu;
 import exception.ErrorException;
 import exception.ExceptionHandler;
 import java.io.IOException;
-import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.util.EnumMap;
 import java.util.Map;
 import util.Convertor;
+import util.Evaluator;
 import view.InputView;
 
 public class AttendanceController {
@@ -79,24 +79,10 @@ public class AttendanceController {
     }
 
     private void validateOpenDate(LocalDate attendDate) {
-        if (isWeekend(attendDate.getDayOfWeek()) || isHoliday(attendDate)) {
+        if (Evaluator.isOpenDate(attendDate)) {
             throw new ErrorException(
                     String.format("%d월 %d일 %s요일은 등교일이 아닙니다.", attendDate.getMonthValue(), attendDate.getDayOfMonth(),
                             Convertor.convertDayOfWeekToKorean(attendDate.getDayOfWeek())));
         }
-    }
-
-    private boolean isWeekend(DayOfWeek dayOfWeek) {
-        if (dayOfWeek == DayOfWeek.SATURDAY || dayOfWeek == DayOfWeek.SUNDAY) {
-            return true;
-        }
-        return false;
-    }
-
-    private boolean isHoliday(LocalDate date) {
-        if (date.equals(LocalDate.of(2024, 12, 25))) {
-            return true;
-        }
-        return false;
     }
 }
